Add files via upload
For a retail furniture store, predicting future sales is critical to avoiding inventory issues like overstocking or under-stocking. The challenge lies in utilizing time series data from the superstore dataset to forecast furniture sales for the next year accurately. This predictive insight ensures an optimal customer experience, avoids losses, and maintains store sustainability.

Conclusion
We observe that sales of furniture produces seasonal pattern. Early of the year is the off season for furniture sales in the superstore. The sales for furniture increases linearly over time.